+++ packages/manager/apps/container/src/index.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import 'piano-analytics-js/dist/browser/piano-analytics.js';
 import { createRoot } from 'react-dom/client';
 import { HashRouter } from 'react-router-dom';