+++ packages/manager/apps/web/client/app/app.module.js
@@ -299,7 +299,7 @@ export default async (containerEl, shellClient) => {
       },
     ])
     .config(async () => {
-      await shellClient.tracking.setConfig(TRACKING);
+      await shellClient.tracking.setConfig(environment.getRegion(), TRACKING);
     })
     .config(
       /* @ngInject */ (atInternetConfigurationProvider) => {