+++ packages/manager/modules/pci/src/components/project/storages/databases/queryStatistics.class.js
@@ -19,10 +19,5 @@ export default class QueryStatistics {
       stddevTime,
       totalTime,
     });
-    this.minTime = this.minTime?.toFixed(1);
-    this.maxTime = this.maxTime?.toFixed(1);
-    this.stddevTime = this.stddevTime?.toFixed(1);
-    this.totalTime = this.totalTime?.toFixed(1);
-    this.meanTime = this.meanTime?.toFixed(1);
   }
 }