+++ packages/manager/modules/pci/src/projects/project/ai-dashboard/components/ai-guide/ai-guide.component.js
@@ -1,0 +1,8 @@
+import template from './ai-guide.html';
+
+export default {
+  bindings: {
+    guide: '<',
+  },
+  template,
+};