+++ packages/manager/modules/pci/src/projects/project/ai-dashboard/tabs/users-tokens/create-user/create-user.routing.js
@@ -1,0 +1,17 @@
+export default /* @ngInject */ ($stateProvider) => {
+  $stateProvider.state(
+    'pci.projects.project.ai-dashboard.users-tokens.create-user',
+    {
+      url: '/create-user',
+      views: {
+        modal: {
+          component: 'pciProjectAiDashboardCreateUserModal',
+        },
+      },
+      layout: 'modal',
+      resolve: {
+        goBack: /* @ngInject */ (goBackToUsersToken) => goBackToUsersToken,
+      },
+    },
+  );
+};