+++ packages/manager/apps/cda/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.7.9](https://github.com/ovh/manager/compare/@ovh-ux/manager-cda-app@1.7.8...@ovh-ux/manager-cda-app@1.7.9) (2023-03-06)
+
+**Note:** Version bump only for package @ovh-ux/manager-cda-app
+
+
+
+
+
 ## [1.7.8](https://github.com/ovh/manager/compare/@ovh-ux/manager-cda-app@1.7.7...@ovh-ux/manager-cda-app@1.7.8) (2023-03-01)
 
 **Note:** Version bump only for package @ovh-ux/manager-cda-app