+++ packages/manager/apps/dbaas-logs/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.10.1](https://github.com/ovh/manager/compare/@ovh-ux/manager-dbaas-logs-app@1.10.0...@ovh-ux/manager-dbaas-logs-app@1.10.1) (2023-03-06)
+
+**Note:** Version bump only for package @ovh-ux/manager-dbaas-logs-app
+
+
+
+
+
 # [1.10.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-dbaas-logs-app@1.9.9...@ovh-ux/manager-dbaas-logs-app@1.10.0) (2023-03-01)